ci: pipeline caching and execution performance (#63)
The purpose of this Pull Request is to improve the execution time of the
PR/CI pipelines. This is primarily achieved using caching.

- Remove the use of `devcontainer/ci` in PR - It provided little to no
value, slowed down execution time, and did not support caching.
- Build and push `devcontainer/ci` in CI - publishing a CI image that
can be used, as required. Mitigating the need to build an image every
workflow run.
- Enabling caching for `pre-commit` - removing the need to initialise
the environment every time.
- Enabling caching for `poetry` - removing the need to `install` deps
every time

Time savings is approximately 3 minutes and 30 seconds per run.
 
Before:

![image](https://github.com/user-attachments/assets/f97b8183-7234-41c6-8667-43f8b7bc531c)

After:

![image](https://github.com/user-attachments/assets/a7303ea6-ed40-4be9-a25d-130b78bba593)

---------

Co-authored-by: Liam Moat <liam.moat@microsoft.com>