feat: Update OAIEmbedding class to support custom embedding dimensions.  (#81)
This commit modifies the OAIEmbedding class in oai.py to support custom
embedding dimensions. The code now checks for the presence of the
EMBEDDING_DIMENSION environment variable and uses it to determine the
number of dimensions for the output embeddings. If the variable is not
set, the code falls back to the default behavior of using the model's
default embedding dimensions.

---------

Co-authored-by: Hilik Berezin <hilikbe@microsoft.com>