refactor: Consolidate configuration (#95)
1. In `experiment.yml` the inference flow is called with the data source
`data/bert-paper-qna-1-line.jsonl`.

Previously, the deployment name was hardcoded in a few places.
- I have updated the inference flow to have a `deployment_name` input,
which has a default if there is none provided
- I have updated the `experiment.yml` to use the
`CHAT_MODEL_DEPLOYMENT_NAME` which comes from the data
(`data/bert-paper-qna-1-line.jsonl`).

2. Previously, the deployment name and other fields were contained
within a `config` object in the data file. It is possible to pass an
object as an input in the flow.dag.yaml, e.g.

```
inputs:  
  config:
    type: object
```
However I could not find a way (or any examples) of how to access the
properties within the object from the flow. For example
`${inputs.config.deployment_name}` wouldn't work.

Instead, I opted to remove the "config" from the data file, and have the
properties at the top level as you can see in bert-paper-qna-1 and
bert-paper-qna-3.

3. There are some properties in the data files that are not used
anywhere (unless I've missed something). For example, the prompt token
limit. We could include this in the flow as an input, or consider using
more of these properties in another update.

---------

Co-authored-by: Martyna Marcinkowska <martynam@microsoft.com>
Co-authored-by: Virginie Marelli <virginie.x.marelli@gsk.com>
Co-authored-by: Martyna Marcinkowska <martyna.x.marcinkowska@gsk.com>
Co-authored-by: Hilik Berlin <hilik.x.berlin@gsk.com>
Co-authored-by: Hilik Berezin <hilikbe@microsoft.com>
Co-authored-by: Liam Moat <liam.moat@microsoft.com>
Co-authored-by: Frances Tibble <frtibble@microsoft.com>
Co-authored-by: Liam Moat <liam.x.moat@gsk.com>