fix: Applies fix to check out triggering workflow code (#121)
This fixes a bug, which would appear when a CI workflow (#1) would be
triggered and complete. If (shortly after) a CI workflow (#2) would be
triggered and complete quickly, then the current CD workflow would
deploy the code from #2. i.e. the CD workflow always deploys the latest
from main, whatever that may be.

This fixes the issue by checking out the code from the triggering
workflow_run or workflow_dispatch.

Co-authored-by: Frances Tibble <frtibble@microsoft.com>