ci: Allow setting environment manually or using the DEFAULT_ENVIRONMENT variable during PR/CI/CD (#143)
# Description
This PR update how the PR, CI and CD workflows work with TMP_ENV
repository variable that used to control the environment that
deployment/evaluation was run on. instead the following changes are
proposed:

- TMP_ENV is deleted and not used anymore
- Each workflow (PR/ CI/ CD) can be either run automatically or
manually.
- When run automatically, the default environment is used. the default
environment is set using a new repository variable 'DEFAULT_ENVRIONMENT'
- When run manually, the user can either override the environment by
setting the dispatch_workflow input parameter 'environment', or leave it
empty in which case the workflow will default back to the
'DEFAULT_ENVRIONMENT' repository variable.

Other changes that are part of this PR
- inverted the dependency between CI and CD workflows. CI triggers CD
with a workflow_dispatch event instead of running CD with a workflow_run
event when the CI completes. This allows the CD to have the same
'environment' input as the CI does.
- removed the dependency between CI and create-release to allow running
the CI without it affecting the release cycle.
- 
* ...

## Does this introduce a breaking change?
<!-- Mark one with an "x". -->

-[ ] Yes
-[X] No

<!-- Please prefix your PR title with one of the following:
  * `feat`: A new feature
  * `fix`: A bug fix
  * `docs`: Documentation only changes
* `style`: Changes that do not affect the meaning of the code
(white-space, formatting, missing semi-colons, etc)
* `refactor`: A code change that neither fixes a bug nor adds a feature
  * `perf`: A code change that improves performance
  * `test`: Adding missing tests or correcting existing tests
* `build`: Changes that affect the build system or external dependencies
(example scopes: gulp, broccoli, npm)
* `ci`: Changes to our CI configuration files and scripts (example
scopes: Travis, Circle, BrowserStack, SauceLabs)
  * `chore`: Other changes that don't modify src or test files
  * `revert`: Reverts a previous commit
* !: A breaking change is indicated with a `!` after the listed prefixes
above, e.g. `feat!`, `fix!`, `refactor!`, etc.
-->

## How to Test
run PR, CI and CD with and without setting the 'environment' variable
and see that they complete successfully on the desired environment.